requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-cors-headers==3.9.0
djangorestframework==3.12.4
drf-nested-routers==0.93.3
idna==3.2
pkg_resources==0.0.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
